+++ yeptap/Autoloader.php
@@ -14,6 +14,7 @@ class Autoloader
                             __DIR__ . DIRECTORY_SEPARATOR,
                             __DIR__ . DIRECTORY_SEPARATOR . 'base',
                             __DIR__ . DIRECTORY_SEPARATOR . 'helpers',
+                            __DIR__ . DIRECTORY_SEPARATOR . 'base' . DIRECTORY_SEPARATOR . 'drivers',
                             SERVER_ROOT . APP_DIR . DIRECTORY_SEPARATOR . 'components',
                             SERVER_ROOT . APP_DIR . DIRECTORY_SEPARATOR . 'models',
                             SERVER_ROOT . APP_DIR . DIRECTORY_SEPARATOR . 'controllers',