+++ yeptap/base/Model.php
@@ -2,16 +2,26 @@
 
 namespace yeptap\base;
 
-class Model
+use \yeptap\base\ORM;
+
+class Model extends ORM
 {
     protected $_model;
+    public $db;
 
     /**
      * Initialization that the object may need before it is used
      */
     public function __construct()
     {
+        $con = new \yeptap\base\drivers\PDO(array(
+            'host' => DB_HOST,
+            'user' => DB_USER,
+            'pwd' => DB_PASSWORD,
+            'dbname' => DB_NAME,
+        ));
 
+        $this->db = $con->connect();
     }
 
     /**