+++ yeptap/base/ORM.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace yeptap\base;
+
+class ORM
+{
+    public function __construct()
+    {
+
+    }
+}